refactor: SelectFormElement を srcType ラッパーパターンに分離
- PlainSelect: 汎用の select 描画コンポーネント
- SrcTypeSelect: srcType 専用のコンテキスト更新ロジック
- SelectFormElement: element.name によるディスパッチャーに変更

TextFormElement と同様のパターンに統一。
